requirements.txt
fastapi
uvicorn
python-dotenv
pymysql
sqlalchemy[mysql]
python-multipart
pydantic
fastapi-mail
